fix int

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -19,7 +19,7 @@ size_t print_listint(const listint_t *h)
 
 	while (ptr != NULL)
 	{
-		printf("%lu", (unsigned long)ptr->n);
+		printf("%d", ptr->n);
 		if (ptr->next == NULL)
 		{
 			printf("\n");
